requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.4
langchain-openai==0.0.5
crewai==0.11.0
pinecone-client==3.0.2
python-dotenv==1.0.1
redis==5.0.1
celery==5.3.6
pytest==8.0.0
httpx==0.26.0 